+++ packages/buefy-next/src/components/carousel/CarouselItem.spec.js
@@ -11,7 +11,7 @@ const WrapperComp = {
         <BCarousel ref="carousel" :animated="animated">
             <BCarouselItem value="item1"/>
             <BCarouselItem ref="testItem" value="item2"/>
-            <BCarouselItem value="item3"/>
+            <BCarouselItem ref="testItem2" value="item3"/>
         </BCarousel>`,
     props: {
         // to indirectly change BCarousel's animated prop
@@ -55,5 +55,45 @@ describe('BCarouselItem', () => {
         await wrapperParent.setProps({ animated: 'fade' })
         wrapperCarousel.vm.changeActive(0)
         expect(wrapper.vm.transition).toBe('fade')
+    })
+
+    it('update active state', () => {
+        expect(wrapper.vm.isActive).toBe(false)
+
+        wrapperCarousel.vm.changeActive(1)
+        expect(wrapper.vm.isActive).toBe(true)
+
+        wrapperCarousel.vm.changeActive(2)
+        expect(wrapper.vm.isActive).toBe(false)
+    })
+
+    describe('with explicit order', () => {
+        let wrapper
+        let wrapperCarousel
+        let wrapperTestItem
+
+        beforeEach(() => {
+            wrapper = mount({
+                template: `
+                    <BCarousel ref="carousel">
+                        <BCarouselItem ref="testItem" :order="2" />
+                        <BCarouselItem :order="1" />
+                        <BCarouselItem :order="3" />
+                    </BCarousel>`,
+                components: { BCarousel, BCarouselItem }
+            })
+            wrapperCarousel = wrapper.findComponent({ ref: 'carousel' })
+            wrapperTestItem = wrapper.findComponent({ ref: 'testItem' })
+        })
+
+        it('update active state', () => {
+            expect(wrapperTestItem.vm.isActive).toBe(false)
+
+            wrapperCarousel.vm.changeActive(1)
+            expect(wrapperTestItem.vm.isActive).toBe(true)
+
+            wrapperCarousel.vm.changeActive(2)
+            expect(wrapperTestItem.vm.isActive).toBe(false)
+        })
     })
 })