+++ packages/buefy-next/src/utils/make-unique-id.js
@@ -1,5 +1,0 @@
-export function makeUniqueId() {
-    const values = new Uint8Array(12)
-    window.crypto.getRandomValues(values)
-    return Array.prototype.map.call(values, (v) => v.toString(16)).join('')
-}