+++ packages/docs/src/pages/components/carousel/examples/ExFull.vue
@@ -18,10 +18,18 @@
                 <div class="control">
                     <b-switch v-model="repeat" :disabled="!autoPlay">Repeat</b-switch>
                 </div>
+                <div class="control">
+                    <b-switch
+                        :model-value="carousels[3].isHidden"
+                        @update:model-value="carousels[3].isHidden = !carousels[3].isHidden"
+                    >
+                        Hide Slide 4
+                    </b-switch>
+                </div>
             </b-field>
             <b-field grouped group-multiline>
                 <b-field label="Value">
-                    <b-numberinput v-model="carousel" min="0" :max="carousels.length - 1" controls-position="compact"/>
+                    <b-numberinput v-model="carousel" min="0" :max="visibleCarousels.length - 1" controls-position="compact"/>
                 </b-field>
                 <b-field label="Interval">
                     <b-numberinput v-model="interval" min="0" controls-position="compact" step="1000" :disabled="!autoPlay"/>
@@ -59,7 +67,11 @@
             :interval="interval"
             :repeat="repeat"
             @change="info($event)">
-            <b-carousel-item v-for="(carousel, i) in carousels" :key="i">
+            <b-carousel-item
+                v-for="(carousel, i) in visibleCarousels"
+                :key="carousel.order"
+                :order="i"
+            >
                 <section :class="`hero is-medium is-${carousel.color} is-bold`">
                     <div class="hero-body has-text-centered">
                         <h1 class="title">{{carousel.title}}</h1>
@@ -89,10 +101,15 @@ export default {
                 { title: 'Slide 1', color: 'dark' },
                 { title: 'Slide 2', color: 'primary' },
                 { title: 'Slide 3', color: 'info' },
-                { title: 'Slide 4', color: 'success' },
+                { title: 'Slide 4', color: 'success', isHidden: false },
                 { title: 'Slide 5', color: 'warning' },
                 { title: 'Slide 6', color: 'danger' }
             ]
+        }
+    },
+    computed: {
+        visibleCarousels() {
+            return this.carousels.filter(carousel => !carousel.isHidden)
         }
     },
     methods: {