+++ CHANGELOG.md
@@ -12,6 +12,10 @@
   If `true`, they are applied to the root `<div>` element, which is compatible with Buefy for Vue 2.
   The default value can be controlled by the `defaultCompatFallthrough` config option (`true` by default).
   [#16](https://github.com/ntohq/buefy-next/issues/16)
+* `CarouselItem`, `StepItem`, and `TabItem` introduce a new prop `order`, which determines the order of each child item.
+  By default, the order of each child item is determined by the sequence in which each child item is mounted.
+  If any child item is unmounted and mounted again, the order may be changed (**BREAKING CHANGE**).
+  You have to give explicit `order` if you want to maintain the order of child items consistent.
 
 ## buefy-next